(mvp) base line refactor
