+++ src/applications/check-in/day-of/components/AppointmentDisplay/AppointmentAction.jsx
@@ -1,18 +1,21 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useState, useMemo } from 'react';
 import PropTypes from 'prop-types';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { api } from '../../../api';
 
-import { URLS } from '../../../utils/navigation/day-of';
 import { useFormRouting } from '../../../hooks/useFormRouting';
 import { ELIGIBILITY, areEqual } from '../../../utils/appointment/eligibility';
 import recordEvent from 'platform/monitoring/record-event';
 import format from 'date-fns/format';
 
+import { makeSelectForm } from '../../../selectors';
+
 import { appointmentWasCheckedInto } from '../../../actions/day-of';
 
 const AppointmentAction = props => {
   const { appointment, router, token } = props;
+  const selectForm = useMemo(makeSelectForm, []);
+  const { urls } = useSelector(selectForm);
 
   const [isCheckingIn, setIsCheckingIn] = useState(false);
   const dispatch = useDispatch();
@@ -23,9 +26,11 @@ const AppointmentAction = props => {
     [dispatch],
   );
 
+  // select URLS from redux using useselector
+
   const defaultMessage =
     'Online check-in isn’t available for this appointment. Check in with a staff member.';
-  const { goToNextPage, goToErrorPage } = useFormRouting(router, URLS);
+  const { goToNextPage, goToErrorPage } = useFormRouting(router, urls);
   const onClick = async () => {
     recordEvent({
       event: 'cta-button-click',
