+++ src/applications/check-in/day-of/components/BackToAppointments.jsx
@@ -1,22 +1,27 @@
-import React from 'react';
+import React, { useMemo } from 'react';
+import { useSelector } from 'react-redux';
+
 import PropTypes from 'prop-types';
 
 import recordEvent from 'platform/monitoring/record-event';
 
 import { createAnalyticsSlug } from '../../utils/analytics';
-import { URLS } from '../../utils/navigation/day-of';
 import { useFormRouting } from '../../hooks/useFormRouting';
 import { withRouter } from 'react-router';
 
+import { makeSelectForm } from '../../selectors';
+
 const BackToAppointments = ({ router, triggerRefresh }) => {
-  const { jumpToPage } = useFormRouting(router, URLS);
+  const selectForm = useMemo(makeSelectForm, []);
+  const { urls } = useSelector(selectForm);
+  const { jumpToPage } = useFormRouting(router, urls);
   const handleClick = e => {
     e.preventDefault();
     recordEvent({
       event: createAnalyticsSlug('back-button-clicked'),
     });
     triggerRefresh();
-    jumpToPage(URLS.DETAILS);
+    jumpToPage(urls.DETAILS);
   };
   return (
     <>
