+++ src/applications/check-in/day-of/pages/Demographics.jsx
@@ -2,7 +2,6 @@ import React, { useCallback, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import recordEvent from 'platform/monitoring/record-event';
-import { URLS } from '../../utils/navigation/day-of';
 import { useFormRouting } from '../../hooks/useFormRouting';
 import BackToHome from '../components/BackToHome';
 import Footer from '../components/Footer';
@@ -13,7 +12,7 @@ import { makeSelectDemographicData } from '../hooks/selectors';
 const Demographics = props => {
   const selectDemographicData = useMemo(makeSelectDemographicData, []);
   const { demographics } = useSelector(selectDemographicData);
-  const { router } = props;
+  const { router, URLS } = props;
   const { goToNextPage, jumpToPage, goToErrorPage } = useFormRouting(
     router,
     URLS,
@@ -79,6 +78,7 @@ const Demographics = props => {
 
 Demographics.propTypes = {
   router: PropTypes.object,
+  URLS: PropTypes.object,
 };
 
 export default Demographics;
