+++ src/applications/check-in/day-of/pages/UpdateInformationQuestion.jsx
@@ -4,7 +4,6 @@ import PropTypes from 'prop-types';
 import recordEvent from 'platform/monitoring/record-event';
 import { focusElement } from 'platform/utilities/ui';
 
-import { URLS } from '../../utils/navigation/day-of';
 import { useFormRouting } from '../../hooks/useFormRouting';
 import BackToHome from '../components/BackToHome';
 import Footer from '../components/Footer';
@@ -13,7 +12,7 @@ const UpdateInformationQuestion = props => {
   useEffect(() => {
     focusElement('h1');
   }, []);
-  const { router } = props;
+  const { router, URLS } = props;
   const { jumpToPage, goToNextPage } = useFormRouting(router, URLS);
 
   const noButtonClicked = () => {
@@ -66,6 +65,7 @@ const UpdateInformationQuestion = props => {
 
 UpdateInformationQuestion.propTypes = {
   router: PropTypes.object,
+  URLS: PropTypes.object,
 };
 
 export default UpdateInformationQuestion;
