+++ src/applications/check-in/day-of/routes.jsx
@@ -16,7 +16,7 @@ import LoadingPage from './pages/LoadingPage';
 import withFeatureFlip from './containers/withFeatureFlip';
 import withSession from './containers/withSession';
 import withToken from './containers/withToken';
-import withForm from './containers/withForm';
+import withForm from '../containers/withForm';
 import { URLS } from '../utils/navigation/day-of';
 
 const routes = [
@@ -101,7 +101,7 @@ const createRoutesWithStore = () => {
         if (route.permissions) {
           const { requiresForm } = route.permissions;
           if (requiresForm) {
-            component = withForm(component);
+            component = withForm(component, URLS);
           }
         }
